requirements.txt
google-genai==1.27.0
pandas==2.3.1
pydantic==2.11.7
streamlit==1.47.1
tqdm==4.67.1
python-dotenv
pydantic
langchain
langchain-google-genai